fix: simply reload named during DNS ACME challenge
We only need the zone files to be reloaded, which is exactly what rndc
reload does, and that is the command runned by systemd when calling
reload on the named service.

We use the systemd command rather that the rndc one directly as it loags
the reload and it may be extended in the future.